Determine whether a word is positive or negative in IsItLoveYet.
IsItLoveYet represents the problem domain so it should be determining whether a word is positive or negative. The IsItLoveViewController is responsible for receiving actions from the views, determining which messages to send to which objects and updating the views with those responses. It would be nice to have a way where the model changes and the views update without need of the controller. I'll have to see what tricks these Apple frameworks have up their sleeves to make that happen.