Moving updateProgress into IsItLoveYet
I realized while I writing the previous message that the
IsItLoveViewController has no business no anything about whether a word
is positive or negative. It should simply pass a word to the domain.
The domain then determines how to respond. This got me to thinking about
how I could just pass the state of the view's fields into the domain and
have the domain respond accordingly. I think this would akin to react
and elm but I'm a little fuzzy on how exactly that works.