Update progress based on the selected word rather.
I want to update progress based on the selected because that allows
IsItLoveYet to determine whether the selected word is a positive or
negative word. That decision seems to be outside the scope of the
controller since it's a decision about the value of an entity(the word)
that was selected by the user. The word is merely an string passed from
the view to the controller. Once that word is in the context of the
domain meaning erupts from it. In that context it has either positive
or negative connotations. From there the effect on the progress towards
love can be determined by the domain.