Make it possible to select a random type of word.
I want to make the app a little more interesting by having a random word appear each time a word is selected. I also want to make it easy to test the UI so I added in the ability to set the word selection strategy. I need to add something to control that configuration before I can use it in the app, otherwise the tests will break.