Allow the application to use random words and tests to use matching
The UITests need to have more predictable data when running so I made it so the tests can switch the behavior of the application if needed. This took a bit of digging to figure out how to do.

In the end I like the concept of modifying the application environment. It fits much more closely with my ingrained notion of setting environment variables in containers that I do as a web developer.